routes secured with is admin

+++ pizza/src/app/api/categories/route.js
@@ -1,11 +1,17 @@
 import {Category} from "@/models/Category";
 import mongoose from "mongoose";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req){
   mongoose.connect(process.env.MONGO_URL);
   const {name} = await req.json();
-  const categoryDoc = await Category.create({name});
-  return Response.json(categoryDoc);
+  if (await isAdmin()) {
+    const categoryDoc = await Category.create({name});
+    return Response.json(categoryDoc);
+
+  } else {
+    return Response.json({});
+  }
 
 }
 
@@ -13,11 +19,13 @@ export async function POST(req){
 export async function PUT(req){
   mongoose.connect(process.env.MONGO_URL);
   const {_id, name} = await req.json();
-  await Category.updateOne({_id}, {name});
+  if (await isAdmin()) {
+    await Category.updateOne({_id}, {name});
+  }
   return Response.json(true);
 }
 
-export async function GET(req){
+export async function GET( ){
   mongoose.connect(process.env.MONGO_URL);
   return Response.json(
     await Category.find().lean()
@@ -29,7 +37,9 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get('_id');
-  await Category.deleteOne({_id});
-  console.log(_id);
+  if (await isAdmin()) {
+    await Category.deleteOne({_id});
+  }
+  // console.log(_id);
   return Response.json(true);
 }
